[Feature] 상품 옵션 기반 예약 흐름 도입

+++ apps/application/src/main/kotlin/com/reservation/application/payment/YPointPaymentStrategy.kt
@@ -1,11 +1,10 @@
 package com.reservation.application.payment
 
+import com.reservation.application.payment.command.PaymentCommand
 import com.reservation.application.user.UserPointService
 import com.reservation.domain.payment.CancelResult
-import com.reservation.domain.payment.PaymentCommand
 import com.reservation.domain.payment.PaymentExecutionResult
 import com.reservation.domain.payment.PaymentMethod
-import com.reservation.domain.payment.PaymentStrategy
 import org.springframework.stereotype.Component
 
 @Component
@@ -15,7 +14,7 @@ class YPointPaymentStrategy(
     override val method: PaymentMethod = PaymentMethod.Y_POINT
 
     override fun pay(command: PaymentCommand): PaymentExecutionResult {
-        val txId = userPointService.deduct(command.userId, command.amount)
+        val txId = userPointService.deduct(command.userId, command.amount.toLong())
         return PaymentExecutionResult(method = method, amount = command.amount, transactionId = txId)
     }
